remove IncompleteFunction in ScriptCompiler
This was a workaround to handle some cases of using declarations.
Today this workaround is almost no longer needed except for some
peculiar cases for which a better solution should be found, but this
can wait.
Removing IncompleteFunction is a first step to simplify the compilation
process and make it more maintainable.